# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI 内容创作平台后端 API。包含用户注册登录、文章多智能体生成、SSE 进度推送、配额管理等。",
        "title": "AI Content Creator API",
        "termsOfService": "https://github.com/ai-content-creator",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ai-content-creator"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8123",
    "basePath": "/api",
    "paths": {
        "/article/ai-modify-outline": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "根据用户自然语言修改意见，调用 LLM 重新生成大纲。返回修改后的大纲，需再调用 confirm-outline 确认。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "AI 修改大纲",
                "parameters": [
                    {
                        "description": "修改建议",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AiModifyOutlineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/article/confirm-outline": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "用户确认或编辑后的大纲，确认后异步进入 Phase3 生成正文与配图，进度通过 SSE 推送。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "确认大纲",
                "parameters": [
                    {
                        "description": "确认大纲参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ConfirmOutlineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/article/confirm-title": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "用户从标题方案中选择主副标题，可选填写补充描述。确认后异步进入 Phase2 生成大纲，进度通过 SSE 推送。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "确认标题",
                "parameters": [
                    {
                        "description": "确认标题参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ConfirmTitleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/article/create": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "创建 AI 文章生成任务，扣减配额后异步生成标题方案。返回 taskId，前端通过 SSE 订阅进度。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "创建文章任务",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误（code 非 0，如参数错误、未登录、配额不足）",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/article/delete": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "逻辑删除文章（isDelete=1）。普通用户只能删除自己的文章，管理员可删除任意文章。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "description": "删除参数（文章主键 id，非 taskId）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/article/execution-logs/{taskId}": {
            "get": {
                "description": "获取指定任务的智能体执行统计与各 Agent 详细日志（耗时、状态、Prompt 等）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "获取智能体执行日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID（UUID）",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/article/list": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "分页查询文章列表。非管理员只能查看自己的文章；管理员可按 userId 筛选。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "分页查询文章列表",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/article/progress/{taskId}": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "建立 Server-Sent Events 长连接，实时推送生成进度。需登录且有权访问该任务。响应 Content-Type 为 text/event-stream，事件名 message，data 为 JSON 字符串。",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "SSE 推送文章生成进度",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID（UUID）",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE 事件流",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "无权访问该任务",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/article/{taskId}": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "根据 taskId 获取文章详情，含标题、大纲、正文、状态、阶段等。普通用户只能查看自己的文章，管理员可查看全部。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "获取文章详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID（UUID）",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误（未登录、无权限、任务不存在）",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码，0 表示成功，非 0 表示失败",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，成功时返回；失败时通常为 null"
                },
                "message": {
                    "description": "响应消息，成功时为 ok，失败时为错误描述",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "model.AiModifyOutlineRequest": {
            "type": "object",
            "required": [
                "modifySuggestion",
                "taskId"
            ],
            "properties": {
                "modifySuggestion": {
                    "description": "用户对大纲的修改建议（自然语言）",
                    "type": "string",
                    "example": "增加一节关于 AI 伦理的内容"
                },
                "taskId": {
                    "description": "任务 ID（UUID）",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "model.ConfirmOutlineRequest": {
            "type": "object",
            "required": [
                "outline",
                "taskId"
            ],
            "properties": {
                "outline": {
                    "description": "确认后的大纲章节列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OutlineSection"
                    }
                },
                "taskId": {
                    "description": "任务 ID（UUID）",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "model.ConfirmTitleRequest": {
            "type": "object",
            "required": [
                "selectedMainTitle",
                "selectedSubTitle",
                "taskId"
            ],
            "properties": {
                "selectedMainTitle": {
                    "description": "用户选定的主标题",
                    "type": "string",
                    "example": "AI 重塑未来教育"
                },
                "selectedSubTitle": {
                    "description": "用户选定的副标题",
                    "type": "string",
                    "example": "从技术到课堂的变革之路"
                },
                "taskId": {
                    "description": "任务 ID（UUID）",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "userDescription": {
                    "description": "用户补充描述（可选）",
                    "type": "string",
                    "example": "面向 K12 教师，语气专业"
                }
            }
        },
        "model.CreateArticleRequest": {
            "type": "object",
            "required": [
                "topic"
            ],
            "properties": {
                "enabledImageMethods": {
                    "description": "允许的配图方式，为空表示支持全部",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "PEXELS",
                        "MERMAID"
                    ]
                },
                "style": {
                    "description": "文章风格，可为空则使用默认",
                    "type": "string",
                    "enum": [
                        "tech",
                        "emotional",
                        "educational",
                        "humorous"
                    ],
                    "example": "tech"
                },
                "topic": {
                    "description": "文章主题/关键词（必填）",
                    "type": "string",
                    "example": "AI 如何改变教育"
                }
            }
        },
        "model.DeleteRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "要删除的记录主键 ID（用户或文章）",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "model.OutlineSection": {
            "type": "object",
            "properties": {
                "points": {
                    "description": "章节要点列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "section": {
                    "description": "章节序号",
                    "type": "integer"
                },
                "title": {
                    "description": "章节标题",
                    "type": "string"
                }
            }
        },
...
        "model.QueryArticleRequest": {
            "type": "object",
            "properties": {
                "pageNum": {
                    "description": "页码，从 1 开始",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer",
                    "example": 10
                },
                "status": {
                    "description": "按任务状态筛选（可选）",
                    "type": "string",
                    "enum": [
                        "PENDING",
                        "PROCESSING",
                        "COMPLETED",
                        "FAILED"
                    ],
                    "example": "COMPLETED"
                },
                "userId": {
                    "description": "按用户 ID 筛选（管理员可用，可选）",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "description": "登录成功后服务端写入 Session Cookie，受保护接口需携带",
            "type": "apiKey",
            "name": "session",
            "in": "cookie"
        }
    }
}